requirements.txt
alembic==1.5.3
Babel==2.9.0
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.6.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-Table==0.5.0
Flask-WTF==0.14.3
geojson==2.5.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
